pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
version = "0.0.0"
name = "omero-isa"
authors = [{ name = "Christoph Möhl" }]
description = "Plugin for omero-cli-transfer to export OMERO data to ISA model metadata."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",

]
dependencies = [
    "omero-py",
    "pandas",
    "isatools",
    "omero-cli-transfer",
    "ome-types",
]

[project.entry-points."omero_cli_transfer.pack.plugin"]
isa = "omero_isa:pack_isa"

[project.scripts]
omero-isa = "omero_isa.cli:main"

[project.urls]
Homepage = "https://github.com/cmohl2013/omero-isa"
